use integers as input in populate(...) methods

+++ src/main/java/net/sci/array/Array2D.java
@@ -64,7 +64,7 @@ public abstract class Array2D<T> implements Array<T>
      * {@code
      * Array2D<String> array = Array3D.create(5, 4, null); 
      * String[] digits = {"A", "B", "C", "D", "E", "F"}; 
-     * array.populate((x,y) -> digits[y.intValue()] + digits[x.intValue()]);
+     * array.populate((x,y) -> digits[y] + digits[x]);
      * String res43 = array.get(4, 3); // returns "CD". 
      * }
      * 
@@ -72,11 +72,11 @@ public abstract class Array2D<T> implements Array<T>
      *            a function of two variables that returns an instance of type
      *            T. The two input variables correspond to the x and y coordinates.
      */
-    public void populate(BiFunction<Double,Double,T> fun)
+    public void populate(BiFunction<Integer,Integer,T> fun)
     {
         for (int[] pos : this.positions())
         {
-            this.set(fun.apply((double) pos[0], (double) pos[1]), pos[0], pos[1]);
+            this.set(fun.apply(pos[0], pos[1]), pos[0], pos[1]);
         }
     }
