+++ src/test/java/net/sci/array/scalar/ScalarArray2DTest.java
@@ -37,7 +37,7 @@ public class ScalarArray2DTest
     {
         ScalarArray2D<?> array = UInt8Array2D.create(10, 10);
         
-        array.populateValues((Double[] pos) -> pos[0] + pos[1]);
+        array.populateValues((Integer[] pos) -> (double) pos[0] + pos[1]);
         
         assertEquals(array.getValue(0, 0), 0, .01);
         assertEquals(array.getValue(9, 0), 9, .01);
@@ -53,7 +53,7 @@ public class ScalarArray2DTest
     {
         ScalarArray2D<?> array = UInt8Array2D.create(10, 10);
         
-        array.populateValues((x,y) -> x + y);
+        array.populateValues((x,y) -> (double) x + y);
         
         assertEquals(array.getValue(0, 0), 0, .01);
         assertEquals(array.getValue(9, 0), 9, .01);
